Scale. [PART 6] (floating point \be strength support)

+++ src/subtitles/Rasterizer.h
@@ -221,15 +221,15 @@ public:
 
     static bool Rasterize(const ScanLineData2& scan_line_data2, int xsub, int ysub, SharedPtrOverlay overlay);
     
-    static bool IsItReallyBlur(int be_strength, double gaussian_blur_strength);
-    static bool OldFixedPointBlur(const Overlay& input_overlay, int be_strength, double gaussian_blur_strength, 
+    static bool IsItReallyBlur(float be_strength, double gaussian_blur_strength);
+    static bool OldFixedPointBlur(const Overlay& input_overlay, float be_strength, double gaussian_blur_strength, 
         double target_scale_x, double target_scale_y, SharedPtrOverlay output_overlay);
 
-    static bool Blur(const Overlay& input_overlay, int be_strength, double gaussian_blur_strength, 
+    static bool Blur(const Overlay& input_overlay, float be_strength, double gaussian_blur_strength, 
         double target_scale_x, double target_scale_y, SharedPtrOverlay output_overlay);
 
-    static bool BeBlur(const Overlay& input_overlay, int be_strength, 
-        double target_scale_x, double target_scale_y, SharedPtrOverlay output_overlay);
+    static bool BeBlur(const Overlay& input_overlay, float be_strength, float target_scale_x, float target_scale_y, 
+        SharedPtrOverlay output_overlay);
 
     static bool GaussianBlur( const Overlay& input_overlay, double gaussian_blur_strength, 
         double target_scale_x, double target_scale_y,
